+++ src/CMakeLists.txt
@@ -326,19 +326,19 @@ engine/video/context.h
 engine/video/coord_sys.h
 engine/video/fade.cpp
 engine/video/fade.h
-engine/video/gl/shader.cpp
-engine/video/gl/shader.h
-engine/video/gl/shader_definitions.h
-engine/video/gl/shader_program.cpp
-engine/video/gl/shader_program.h
-engine/video/gl/shader_programs.h
-engine/video/gl/shaders.h
-engine/video/gl/sprite.cpp
-engine/video/gl/sprite.h
-engine/video/gl/sprite_particle_system.cpp
-engine/video/gl/sprite_particle_system.h
-engine/video/gl/transform.cpp
-engine/video/gl/transform.h
+engine/video/gl/gl_particle_system.cpp
+engine/video/gl/gl_particle_system.h
+engine/video/gl/gl_shader.cpp
+engine/video/gl/gl_shader.h
+engine/video/gl/gl_shader_definitions.h
+engine/video/gl/gl_shader_program.cpp
+engine/video/gl/gl_shader_program.h
+engine/video/gl/gl_shader_programs.h
+engine/video/gl/gl_shaders.h
+engine/video/gl/gl_sprite.cpp
+engine/video/gl/gl_sprite.h
+engine/video/gl/gl_transform.cpp
+engine/video/gl/gl_transform.h
 engine/video/image.cpp
 engine/video/image.h
 engine/video/image_base.cpp
