+++ src/engine/video/gl/gl_shader.h
@@ -8,32 +8,37 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 /** ****************************************************************************
-*** \file    shaders.h
+*** \file    gl_shader.h
 *** \author  Authenticate, James Lammlein
-*** \brief   Header file for shader definitions.
+*** \brief   Header file for shaders used in OpenGL.
 *** ***************************************************************************/
 
-#ifndef __SHADERS_HEADER__
-#define __SHADERS_HEADER__
+#ifndef __SHADER_HEADER__
+#define __SHADER_HEADER__
 
 namespace vt_video
 {
 namespace gl
 {
-namespace shaders
-{
 
-enum Shaders
+// Forward declarations.
+class ShaderProgram;
+
+//! \brief A class for an OpenGL shader.
+class Shader
 {
-    VertexDefault = 0,
-    FragmentSolid,
-    FragmentSolidGrayscale,
-    FragmentSprite,
-    FragmentSpriteGrayscale,
-    Count
+    friend class ShaderProgram;
+
+public:
+    Shader(GLenum type, const std::string &data);
+    ~Shader();
+
+private:
+    GLint _shader;
+
+    Shader(const Shader&) {}
+    Shader& operator=(const Shader&) { return *this; }
 };
-
-} // namespace shaders
 
 } // namespace gl
