+++ src/engine/video/gl/gl_shader_definitions.h
@@ -9,7 +9,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 /** ****************************************************************************
-*** \file    shader_definitions.h
+*** \file    gl_shader_definitions.h
 *** \author  Authenticate, James Lammlein
 *** \brief   Header file for shader definitions.
 *** ***************************************************************************/
