+++ src/engine/video/gl/gl_shader_program.cpp
@@ -8,15 +8,15 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 /** ****************************************************************************
-*** \file    shader_program.cpp
+*** \file    gl_shader_program.cpp
 *** \author  Authenticate, James Lammlein
 *** \brief   Source file for shader programs used in OpenGL.
 *** ***************************************************************************/
 
 #include "utils/utils_pch.h"
-#include "shader_program.h"
+#include "gl_shader_program.h"
 
-#include "shader.h"
+#include "gl_shader.h"
 
 #include "utils/utils_strings.h"
