+++ src/engine/video/gl/gl_shader_program.h
@@ -8,7 +8,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 /** ****************************************************************************
-*** \file    shader_program.h
+*** \file    gl_shader_program.h
 *** \author  Authenticate, James Lammlein
 *** \brief   Header file for shader programs used in OpenGL.
 *** ***************************************************************************/
