+++ src/engine/video/gl/gl_shader_programs.h
@@ -8,7 +8,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 /** ****************************************************************************
-*** \file    shader_programs.h
+*** \file    gl_shader_programs.h
 *** \author  Authenticate, James Lammlein
 *** \brief   Header file for shader program definitions.
 *** ***************************************************************************/
