+++ src/engine/video/video.cpp
@@ -21,13 +21,13 @@
 #include "engine/mode_manager.h"
 #include "engine/script/script_read.h"
 #include "engine/system.h"
-#include "engine/video/gl/shader.h"
-#include "engine/video/gl/shader_definitions.h"
-#include "engine/video/gl/shader_program.h"
-#include "engine/video/gl/shader_programs.h"
-#include "engine/video/gl/shaders.h"
-#include "engine/video/gl/sprite.h"
-#include "engine/video/gl/sprite_particle_system.h"
+#include "engine/video/gl/gl_particle_system.h"
+#include "engine/video/gl/gl_shader.h"
+#include "engine/video/gl/gl_shader_definitions.h"
+#include "engine/video/gl/gl_shader_program.h"
+#include "engine/video/gl/gl_shader_programs.h"
+#include "engine/video/gl/gl_shaders.h"
+#include "engine/video/gl/gl_sprite.h"
 
 #include "utils/utils_strings.h"
