+++ src/engine/video/video.h
@@ -33,9 +33,9 @@
 #include "engine/video/context.h"
 #include "engine/video/coord_sys.h"
 #include "engine/video/fade.h"
-#include "engine/video/gl/shader_programs.h"
-#include "engine/video/gl/shaders.h"
-#include "engine/video/gl/transform.h"
+#include "engine/video/gl/gl_shader_programs.h"
+#include "engine/video/gl/gl_shaders.h"
+#include "engine/video/gl/gl_transform.h"
 #include "engine/video/image.h"
 #include "engine/video/screen_rect.h"
 #include "engine/video/text.h"
