+++ vs2013/ValyriaTear/ValyriaTear.vcxproj
@@ -136,11 +136,11 @@
     <ClCompile Include="..\..\src\engine\script_supervisor.cpp" />
     <ClCompile Include="..\..\src\engine\system.cpp" />
     <ClCompile Include="..\..\src\engine\video\fade.cpp" />
-    <ClCompile Include="..\..\src\engine\video\gl\shader.cpp" />
-    <ClCompile Include="..\..\src\engine\video\gl\shader_program.cpp" />
-    <ClCompile Include="..\..\src\engine\video\gl\sprite.cpp" />
-    <ClCompile Include="..\..\src\engine\video\gl\sprite_particle_system.cpp" />
-    <ClCompile Include="..\..\src\engine\video\gl\transform.cpp" />
+    <ClCompile Include="..\..\src\engine\video\gl\gl_particle_system.cpp" />
+    <ClCompile Include="..\..\src\engine\video\gl\gl_shader.cpp" />
+    <ClCompile Include="..\..\src\engine\video\gl\gl_shader_program.cpp" />
+    <ClCompile Include="..\..\src\engine\video\gl\gl_sprite.cpp" />
+    <ClCompile Include="..\..\src\engine\video\gl\gl_transform.cpp" />
     <ClCompile Include="..\..\src\engine\video\image.cpp" />
     <ClCompile Include="..\..\src\engine\video\image_base.cpp" />
     <ClCompile Include="..\..\src\engine\video\interpolator.cpp" />
@@ -229,14 +229,14 @@
     <ClInclude Include="..\..\src\engine\video\context.h" />
     <ClInclude Include="..\..\src\engine\video\coord_sys.h" />
     <ClInclude Include="..\..\src\engine\video\fade.h" />
-    <ClInclude Include="..\..\src\engine\video\gl\shader.h" />
-    <ClInclude Include="..\..\src\engine\video\gl\shaders.h" />
-    <ClInclude Include="..\..\src\engine\video\gl\shader_definitions.h" />
-    <ClInclude Include="..\..\src\engine\video\gl\shader_program.h" />
-    <ClInclude Include="..\..\src\engine\video\gl\shader_programs.h" />
-    <ClInclude Include="..\..\src\engine\video\gl\sprite.h" />
-    <ClInclude Include="..\..\src\engine\video\gl\sprite_particle_system.h" />
-    <ClInclude Include="..\..\src\engine\video\gl\transform.h" />
+    <ClInclude Include="..\..\src\engine\video\gl\gl_particle_system.h" />
+    <ClInclude Include="..\..\src\engine\video\gl\gl_shader.h" />
+    <ClInclude Include="..\..\src\engine\video\gl\gl_shaders.h" />
+    <ClInclude Include="..\..\src\engine\video\gl\gl_shader_definitions.h" />
+    <ClInclude Include="..\..\src\engine\video\gl\gl_shader_program.h" />
+    <ClInclude Include="..\..\src\engine\video\gl\gl_shader_programs.h" />
+    <ClInclude Include="..\..\src\engine\video\gl\gl_sprite.h" />
+    <ClInclude Include="..\..\src\engine\video\gl\gl_transform.h" />
     <ClInclude Include="..\..\src\engine\video\image.h" />
     <ClInclude Include="..\..\src\engine\video\image_base.h" />
     <ClInclude Include="..\..\src\engine\video\interpolator.h" />
